+++ instrumentation/opentelemetry-instrumentation-confluent-kafka/tests/test_instrumentation.py
@@ -23,6 +23,9 @@ from opentelemetry.instrumentation.confluent_kafka import (
     ProxiedConsumer,
     ProxiedProducer,
 )
+from opentelemetry.instrumentation.confluent_kafka.utils import (
+    KafkaContextSetter,
+)
 
 
 class TestConfluentKafka(TestCase):
@@ -73,3 +76,16 @@ class TestConfluentKafka(TestCase):
         consumer = instrumentation.instrument_consumer(consumer)
         self.assertEqual(consumer.__class__, ProxiedConsumer)
         self.assertTrue(hasattr(consumer, "commit"))
+
+    def test_context_setter(self) -> None:
+        context_setter = KafkaContextSetter()
+
+        carrier_dict = {"key1": "val1"}
+        context_setter.set(carrier_dict, "key2", "val2")
+        self.assertGreaterEqual(
+            carrier_dict.items(), {"key2": "val2".encode()}.items()
+        )
+
+        carrier_list = [("key1", "val1")]
+        context_setter.set(carrier_list, "key2", "val2")
+        self.assertTrue(("key2", "val2".encode()) in carrier_list)